+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Change 1.19.7
 
-- Add `longForm,MsBackend` mehod (issue
+- Add `longForm,MsBackend` and `longForm,Spectra` methods (issue
   [#360](https://github.com/rformassspectrometry/Spectra/issues/360)).
 
 ## Change 1.19.6